+++ packages/client/components/RetroReflectPhase/PhaseItemEditor.tsx
@@ -12,6 +12,7 @@ import {useTipTapReflectionEditor} from '../../hooks/useTipTapReflectionEditor'
 import CreateReflectionMutation from '../../mutations/CreateReflectionMutation'
 import EditReflectionMutation from '../../mutations/EditReflectionMutation'
 import {Elevation} from '../../styles/elevation'
+import {PALETTE} from '../../styles/paletteV3'
 import {BezierCurve, ZIndex} from '../../types/constEnums'
 import {cn} from '../../ui/cn'
 import ReflectionCardAuthor from '../ReflectionCard/ReflectionCardAuthor'
@@ -32,6 +33,21 @@ const CardInFlightStyles = styled(ReflectionCardRoot)<{transform: string; isStar
     zIndex: ZIndex.REFLECTION_IN_FLIGHT
   })
 )
+
+const EnterHint = styled('div')<{visible: boolean}>(({visible}) => ({
+  color: PALETTE.SLATE_600,
+  fontSize: 14,
+  fontStyle: 'italic',
+  fontWeight: 400,
+  lineHeight: '20px',
+  paddingLeft: 16,
+  cursor: 'pointer',
+  visibility: visible ? undefined : 'hidden',
+  opacity: visible ? 1 : 0,
+  height: visible ? 28 : 0,
+  overflow: 'hidden',
+  transition: 'height 300ms, opacity 300ms'
+}))
 
 interface Props {
   cardsInFlightRef: MutableRefObject<ReflectColumnCardInFlight[]>
@@ -66,6 +82,13 @@ const PhaseItemEditor = (props: Props) => {
         disableAnonymity
         viewerMeetingMember {
           user {
+            timeline(first: 2, eventTypes: [retroComplete]) {
+              edges {
+                node {
+                  id
+                }
+              }
+            }
             preferredName
           }
         }
@@ -145,6 +168,33 @@ const PhaseItemEditor = (props: Props) => {
     }
   }, [idleTimerIdRef])
 
+  const knowsHowToEnter = (viewerMeetingMember?.user?.timeline?.edges?.length ?? 0) > 1
+  const [isFocused, setIsFocused] = useState(false)
+  const [enterHint, setEnterHint] = useState('')
+  const hintTimerRef = useRef<number>()
+  const hintCharacterCountRef = useRef(0)
+  useEffect(() => {
+    const showHint = !knowsHowToEnter && !isEditing && !editor?.isEmpty
+    const characterCount = editor?.storage.characterCount.characters()
+
+    if (characterCount !== hintCharacterCountRef.current) {
+      hintCharacterCountRef.current = characterCount
+      setEnterHint('')
+    }
+
+    if (showHint) {
+      const newEnterHint = isFocused
+        ? 'Press enter to add'
+        : 'Forgot to press enter? Click here to add 👆'
+      hintTimerRef.current = window.setTimeout(() => setEnterHint(newEnterHint), 1000)
+      return () => {
+        window.clearTimeout(hintTimerRef.current)
+      }
+    } else {
+      return undefined
+    }
+  }, [isFocused, isEditing, editor?.storage.characterCount.characters()])
+
   const ensureNotEditing = () => {
     if (!isEditing) return
     window.clearTimeout(idleTimerIdRef.current)
@@ -173,10 +223,12 @@ const PhaseItemEditor = (props: Props) => {
     }, 5000)
   }
   const onFocus = () => {
+    setIsFocused(true)
     ensureEditing()
     return null
   }
   const onBlur = () => {
+    setIsFocused(false)
     ensureNotEditing()
   }
 
@@ -210,6 +262,9 @@ const PhaseItemEditor = (props: Props) => {
             <ReflectionCardAuthor>{viewerMeetingMember?.user.preferredName}</ReflectionCardAuthor>
           </div>
         )}
+        <EnterHint visible={!!enterHint} onClick={handleSubmit}>
+          {enterHint}
+        </EnterHint>
       </ReflectionCardRoot>
       {portal(
         <>