+++ packages/server/integrations/mattermost/mattermostWebhookHandler.ts
@@ -1,30 +1,34 @@
-import {createVerifier, httpbis} from 'http-message-signatures'
 import {HttpRequest, HttpResponse} from 'uWebSockets.js'
 import appOrigin from '../../appOrigin'
-import AuthToken from '../../database/types/AuthToken'
 import uWSAsyncHandler from '../../graphql/uWSAsyncHandler'
 import parseBody from '../../parseBody'
-import getKysely from '../../postgres/getKysely'
-import encodeAuthToken from '../../utils/encodeAuthToken'
+import publishWebhookGQL from '../../utils/publishWebhookGQL'
+import sendToSentry from '../../utils/sendToSentry'
 
-const MATTERMOST_SECRET = process.env.MATTERMOST_SECRET
+const login = async (variables: {request: string; body: string}) => {
+  const query = `
+    mutation LoginMattermost($request: String!, $body: String!) {
+      loginMattermost(request: $request, body: $body) {
+        error {
+          message
+        }
+        authToken
+      }
+    }
+  `
 
-const login = async (email: string) => {
-  const pg = getKysely()
-  const user = await pg
-    .selectFrom('User')
-    .selectAll()
-    .where('email', '=', email)
-    .executeTakeFirstOrThrow()
-  const authToken = new AuthToken({sub: user.id, tms: user.tms})
-  return encodeAuthToken(authToken)
+  const loginResult = await publishWebhookGQL<{
+    data: {
+      loginMattermost: {
+        error: {message: string} | null
+        authToken: string | null
+      }
+    }
+  }>(query, variables)
+  return loginResult?.data?.loginMattermost ?? {authToken: null, error: {message: 'Unknown error'}}
 }
 
 const mattermostWebhookHandler = uWSAsyncHandler(async (res: HttpResponse, req: HttpRequest) => {
-  if (!MATTERMOST_SECRET) {
-    res.writeStatus('404').end()
-    return
-  }
   const headers = {
     'content-type': req.getHeader('content-type'),
     'content-digest': req.getHeader('content-digest'),
@@ -33,40 +37,33 @@ const mattermostWebhookHandler = uWSAsyncHandler(async (res: HttpResponse, req: 
     'signature-input': req.getHeader('signature-input')
   }
 
-  const verified = await httpbis.verifyMessage(
-    {
-      async keyLookup(_: any) {
-        // TODO When we support multiple Parabol - Mattermost connections, we should look up the key from IntegrationProvider
-        // const keyId = params.keyid;
-        return {
-          id: '',
-          algs: ['hmac-sha256'],
-          verify: createVerifier(MATTERMOST_SECRET, 'hmac-sha256')
-        }
-      }
-    },
-    {
-      method: req.getMethod(),
-      url: appOrigin + req.getUrl(),
-      headers
-    }
-  )
-  if (!verified) {
-    res.writeStatus('401').end()
-    return
-  }
-
-  const body = await parseBody<{email: string}>({
-    res
+  const request = JSON.stringify({
+    method: req.getMethod(),
+    url: appOrigin + req.getUrl(),
+    headers
   })
 
-  const {email} = body ?? {}
-  if (!email) {
-    res.writeStatus('401').end()
+  const body = await parseBody<string>({
+    res,
+    parser: (buffer) => buffer.toString()
+  })
+
+  if (!body) {
+    res.writeStatus('400').end('Bad Request')
     return
   }
 
-  const authToken = await login(email)
+  const {authToken, error} = await login({
+    request,
+    body
+  })
+
+  if (error) {
+    sendToSentry(new Error(error.message))
+    res.writeStatus('500').end(error.message)
+    return
+  }
+
   res
     .writeStatus('200')
     .writeHeader('Content-Type', 'application/json')